requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
groq
python-jose==3.3.0
bcrypt==4.0.1
gunicorn==21.2.0
